+++ Client/MapleStory2/src/game_object/map/cube/map_object.cpp
@@ -82,7 +82,7 @@ auto MapObject::AddComponents(MapParser::MapEntity& entity) -> HRESULT
 	transformDesc.speed_per_sec = 5.0f;
 	transformDesc.rotation_per_sec = D3DXToRadian(90.0);
 
-	if (FAILED(AddComponent(static_cast<int32_t>(kScene::kSceneStatic), TEXT("Prototype_Transform"), TEXT("Com_Transform"), reinterpret_cast<std::shared_ptr<Component>*>(&_transform_com), &transformDesc)))
+	if (FAILED(AddComponent(kScene::kSceneStatic, TEXT("Prototype_Transform"), TEXT("Com_Transform"), reinterpret_cast<std::shared_ptr<Component>*>(&_transform_com), &transformDesc)))
 		return E_FAIL;
 	if (!entity.propertise.empty())
 	{
